Moved the `engine_client` into the bria_engine_api sub-directory
