+++ src/bria_internal/common/bria_engine_api/image_editing/__init__.py
@@ -1,0 +1,3 @@
+from .background_remove import BackgroundRemoveAPI
+
+__all__ = ["BackgroundRemoveAPI"]
