+++ src/bria_internal/common/bria_engine_api/status/status.py
@@ -4,8 +4,8 @@ from typing import Awaitable
 
 from httpx import Response
 
-from bria_internal.common.bria_engine_api import BriaEngineClient
 from bria_internal.common.bria_engine_api.enable_sync_decorator import enable_run_synchronously
+from bria_internal.common.bria_engine_api.engine_client import BriaEngineClient
 from bria_internal.common.bria_engine_api.routes_constants import BriaEngineAPIRoutes
 from bria_internal.schemas.status_api import StatusAPIResponse, StatusAPIState
